+++ src/app/cli/src/services/config/models.rs
@@ -1183,6 +1183,7 @@ impl Default for EmbeddingsChunkerConfigSimple {
 #[serde(tag = "kind")]
 pub enum EmbeddingsEncoderConfig {
     OpenAi(EmbeddingsEncoderConfigOpenAi),
+    Dummy,
 }
 
 impl Default for EmbeddingsEncoderConfig {