+++ src/domain/search/domain/src/domain/search_schema.rs
@@ -16,6 +16,7 @@ pub type SearchFieldPath = &'static str;
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 pub mod fields {
+    pub const ID: &str = "_id";
     pub const TITLE: &str = "title";
     pub const IS_BANNED: &str = "is_banned";
     pub const SEMANTIC_EMBEDDINGS: &str = "semantic_embeddings";
@@ -39,17 +40,16 @@ pub struct SearchEntitySchema {
     pub upgrade_mode: SearchEntitySchemaUpgradeMode,
     pub fields: &'static [SearchSchemaField],
     pub title_field: SearchFieldPath,
-    pub enable_banning: bool,
+    pub flags: SearchEntitySchemaFlags,
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-impl SearchEntitySchema {
-    pub fn find_embedding_chunks_field(&self) -> Option<&SearchSchemaField> {
-        self.fields
-            .iter()
-            .find(|f| matches!(f.role, SearchSchemaFieldRole::EmbeddingChunks))
-    }
+#[derive(Debug, Clone, Copy)]
+pub struct SearchEntitySchemaFlags {
+    pub enable_banning: bool,
+    pub enable_security: bool,
+    pub enable_embeddings: bool,
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -68,13 +68,14 @@ pub enum SearchSchemaFieldRole {
         enable_inner_ngrams: bool,
     },
     Name,
-    Description,
+    Description {
+        add_keyword: bool,
+    },
     Prose,
     Keyword,
     DateTime,
     Boolean,
     Integer,
-    EmbeddingChunks,
     UnprocessedObject,
     // TODO: Add more field roles as needed
 }